requirements.txt
# ============================================================================
# requirements.txt
# ============================================================================
"""
numpy>=1.24.0,<2.0.0
opencv-python>=4.7.0,<5.0.0
Pillow>=9.5.0,<11.0.0
PyYAML>=6.0,<7.0.0
tqdm>=4.65.0,<5.0.0
pydantic>=2.0.0,<3.0.0
typing-extensions>=4.6.0
bpy>=3.6.0
"""

